Implemented name paser with regex
Scraper can now be programmed to scrape all the urls from a given url, in theory we can range this to get all the urls posing and due to the data having a validation method we can then only examine urls from a certain time, but we are still saving all the urls in memory.

The parser is implemented to only scrape from one url link. There is a strict protocol in place which is creating and missing names.

Further implementation will create a much more flexible regex