Primary goal completed
The program is working as intended. Every link (50 per page) can be scraped. From there we go through each link, during test runs some links did not pull any captions and it is unknown which links did not have captions. The name parser pulls names but it is unsure how accurate this version is. Further testing is needed