Finished product
At this time the program is working as intended.
Future work would include formatting all data that come in initially so that we only need to work with strings instead of unicode. Other recommendations would be to reformat code and create more methods so that code base is easier to read and more adjustable. This is a start that could go in many directions. Social networking, etc.wq
'